requirements.txt
tehblog
django-tagging
-e git://github.com/andrewebdev/django-ostinato.git=django-ostinato
